move redis to src/clients
